adjust test location for nft transfer log
